Added static_cast<size_t> to prevent warning.
git-svn-id: http://svn.purei.org/framework/trunk@267 6a6d099a-6a11-0410-877e-d5a07a98cbd2

+++ include/offsetof_def.h
@@ -2,6 +2,6 @@
 #define _OFFSETOF_DEF_H_
 
 #undef offsetof
-#define offsetof(a, b) (reinterpret_cast<uint8*>(&reinterpret_cast<a*>(0x10)->b) - reinterpret_cast<uint8*>(0x10))
+#define offsetof(a, b) static_cast<size_t>(reinterpret_cast<uint8*>(&reinterpret_cast<a*>(0x10)->b) - reinterpret_cast<uint8*>(0x10))
 
 #endif
